Huge refactor to resolve cyclic dependency + get jcstress to work
