+++ core/src/commonMain/kotlin/komem/litmus/AffinityMap.kt
