+++ core/src/commonMain/kotlin/komem/litmus/LitmusAutoOutcome.kt
