+++ core/src/nativeInterop/kaffinity.def
