+++ core/src/nativeMain/kotlin/komem.litmus/NativeUtils.kt
