+++ jcstress-wrapper/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    kotlin("jvm")
+    application
+}
+
+application {
+    mainClass = "MainKt"
+}
+
+dependencies {
+    implementation(project(":core"))
+    implementation(kotlin("reflect"))
+}
