--- litmus-jcstress/build.gradle.kts
@@ -1,7 +1,0 @@
-plugins {
-    kotlin("jvm")
-}
-
-dependencies {
-    implementation(project(":litmus"))
-}
