--- litmus-jcstress/src/main/kotlin/komem/litmus/CodegenMain.kt
@@ -1,5 +1,0 @@
-package komem.litmus
-
-fun main() {
-    val t = litmusTest
-}
