+++ settings.gradle.kts
@@ -1,5 +1,6 @@
 rootProject.name = "litmuskt"
 
-include(":litmus")
+include(":core")
 include(":codegen")
-include(":litmus-jcstress")
+include(":jcstress-wrapper")
+include(":cli")
